only download if the file is new

+++ src/provor_ftp.py
@@ -1,5 +1,6 @@
 #!/usr/bin/python
 
+import sys
 from pathlib import Path
 import ftplib
 
@@ -12,7 +13,10 @@ for imei in ftp.nlst():
         local_file = Path('../data/provor') / imei / fn
         if not local_file.parent.exists():
             local_file.parent.mkdir()
-        lf = open(local_file, 'wb')
-        ftp.retrbinary('RETR ' + fn, lf.write)
-        lf.close()
+        if not local_file.exists():
+            sys.stdout.write(f'Downloading {imei}/{fn}...')
+            lf = open(local_file, 'wb')
+            ftp.retrbinary('RETR ' + fn, lf.write)
+            lf.close()
+            sys.stdout.write('done\n')
     ftp.cwd('../')
